Error in using "id" in the "configureRateLimiting()" function in the "RouteServiceProvider.php" file.
The code in "RouteServiceProvider.php" file has been modified by replacing "id" with "$request->user()->id ?: $request->ip()" in the "configureRateLimiting()" function to ensure that registered system users are not prevented from accessing the application if they exceed the maximum number of requests from their IP address. The code checks for the existence of a registered user in the system, and if found, uses their ID as a identifier for rate limiting. If no user is found, the IP address of the request is used as the identifier.